Fix crash in stack walking.
Bug: 8537133

The G+ app throws an exception which leads to a crash when walking the stack.
The cause is the compiled code tries to set the top of quick stack while the
current managed stack is shadow frame based.

This CL adds the missing "pop" of the shadow frame and also some checks enabled
in debug only.

Change-Id: Iac04993b287abadee7a3c240e2e9cf1a07e50ce5