Fix ObjectReference.InvokeMethod.
This probably broke when I rewrote the object registry, but because
the test was so crap, we may have gotten away with passing bad pointers.
(Though for me, CheckJNI was catching this.)

While I'm here, fix the argument checking, which was previously
very weak; we'd accept any reference type rather than instances of
the parameter's specific type.

Change-Id: I08c001cabde02a0509fe28df17523a2d2519d1ca