Optimize branch instructions in interpreter.
This CL cleans up branch instructions handling by optimizing instruction
displacement with an offset by changing the way we compute the next
instruction. On ARM, it saves at least 3 instructions for GOTO and 4
instructions for IF-XX/Z.

This CL also transforms the Instruction::At CHECK into a DCHECK. This improves
the compiled code when computing next instruction.

Change-Id: I901699201012d2dbbaef4d9158cc97b63bfbbe84