Support inline dex data
Dx always places switch table and array data following executable
code.  However, embedding data inline appears to be legal - and
the Garmin Connect app does this.  Unknown is what tool generated
the odd dex code.  This CL reworks dex parsing to support inline data.

Further, a verification check to require a move-exception at
the beginning of catch regions is disabled.  Dalvik's verifier appears
to only require that if a move-exception exists, it must be at
the beginning of the catch.  If there is value to the check
requiring a move-exception, we'll need to enable it based on a
future dex version.

Change-Id: I80c78544993acb999f7c62d32479c3c8455b56cb