Fix verifier upcasting type after instance_of.
The verifier automatically changed the type of a register to be
the checked type in an instance_of instruction, even if the
checked type was the register type's superclass. This would loosen
the type information of the register and cause problems later.

Bug: 10614872

(cherry picked from commit c642ec8987746a2a44b990bd5354306242d709da)

Change-Id: Ib447557d3582dad6ac01f0e3b4ee497a27a45172