Set processes' comm names to their package names.
When debugging audit logs, it is often helpful to
have the package's name. Bear in mind that the
package name is the right most chars up to
TASK_COMM_LEN. Although a possibility of truncation
exists, it can be useful for narrowing down the
possibilities.

(cherry-pick of 5ae4516eb30c7cf674025ab01b097e62c7901b98 from AOSP.)

Change-Id: I27dfd986ebe86af4a5b8c702af955d5065daa12f