Remove zygote space from alloc_spaces_ array.
Not removing this space was causing a crash in
Heap::GetObjectsAllocated since the new zygote bin packing places
objects where the normal DlMalloc/RosAlloc accounting would be.

Going to refactor zygote spaces to be a new space type in the near
future to avoid more of these types of issues.

Bug: 12490061

Change-Id: Id4f75d0315c63e0e8c19f6fba9ad9cb2ba9017e5