Run checkpoints after resuming threads in signal catcher SIGQUIT.
Previously we were running checkpoints after restoring the thread
state, but before resuming the threads. This caused an issue with
profiler checkpoints since these use a ScopedObjectAccess. Since
we had with exclusive access to the mutator lock when we ran the
checkpoints, it caused deadlocks when we tried to gain shared
access to the mutator lock in the ScopedThreadStateChange if the
old_state was any state other than kRunnable.

Bug: 12490061
Change-Id: I5f922693a4bf5d18cf9d22f1f28a7a04b6bfc077