Avoid compiling monster methods in boot image
The Quick compiler normally avoids compiling extremely large
methods (which are usually auto-generated initialization
routines) because they take too long to compile, and generally
aren't executed frequently enough to benefit from compilation.

However, we'd previously compiled even large methods in the
boot image - because it is normally cross-compiled on the host.
However, we are currently compiling it on the target in some
situations.  This change moves the default for image creation
from "compile everything" to "compile most things".

Change-Id: If096e1ae5db2e3a394880e36e19607358e77d677