Fix JDWP thread crash.
Fix a crash happenning when JDWP thread attempts to process incoming packets
from a closed socket. It happens when the runtime shuts down while JDWP thread
is still running and waiting for incoming JDWP packets.

When we close the JDWP connection from runtime destructor, we wake up the JDWP
thread. To avoid this, we call Dbg::Dispose before closing the connection.

Bug: 12578041
Change-Id: I8d03478b6c3f5f49e08899d476f292b7a401d5f3