Prevent object allocation related races in VisitObjects.
Prevents the following race conditions:
Someone is in the process of pushing a reference in the allocation
stack but hasn't yet written the reference. This caused VisitObjects
to occasionally send null objects to the visitor.
Fixed another race where the object had just been allocated but the
class had not been written. We now skip objects in the allocation
stack which have a null class.

Bug: 13004631
Change-Id: Iad789c5e277a7717ce595c7124f0d65b44392fd8