Do not set ACC_SUPER flag for interfaces
The debugger's GetModifier method forces the ACC_SUPER flag set for all
classes and interfaces because dex files don't contain this flag.
But according to the JVM Spec (Class File Structure) the ACC_SUPER flag
must not be set for interfaces:
   If the ACC_INTERFACE flag of this class file is set,
   its ACC_ABSTRACT flag must also be set (JLS 9.1.1.1). Such a class
   file must not have its ACC_FINAL, ACC_SUPER or ACC_ENUM flags set.

The patch sets ACC_SUPER only if ACC_INTERFACE is not set.

Signed-off-by: Yevgeny Rouban <yevgeny.y.rouban@intel.com>
Change-Id: I6cc4b215a6584d177845a4f8cce1efeb1650f646