LongToFp result storage fixed
When converting long values to floating-point, the resulting value
was not stored properly. This was the reason of incorrect results
of arithmetic operations where such conversions were used.

Performance measurement in Quadrant Pro KPI benchmark was blocked
by long-to-fp conversion errors.

This patch ensures that result of long-to-fp conversion is stored
properly.

Change-Id: I94beffcf7f0146c0c6d5b8ae33b020c985843db0
Signed-off-by: Maxim Kazantsev <maxim.kazantsev@intel.com>