Quick compiler: Method* as a reference
This is the first of two CLs intended to fix up and make consistent
the handling of references in the Quick backend.  A sibling
CL c/96237 updates the runtime to treat Method* as a compressed
reference when stored.  This CL makes a similar change for the
backend.

As far as the general handling of in-register references, though,
the current Quick backend is not consistent even for non-Method*
references.  Sometimes they are treated as references, but other
times are handled as if they were 32-bit ints.  A subsequent CL
will deal with that issue.

Change-Id: I5591c5eea6cca6ed22208ab806fd38b959c9d03d