ART: Do not attempt to compile a method with too many registers
The mir_graph has a hard limit on how many ssa registers it can
handle, and will silently fail or fatal out in debug builds when
more SSA registers are necessary. This adds a check in the frontend
that tries to catch some of these cases, and makes the DCHECK a
CHECK to not silently fail.

Bug: 16018097

(cherry picked from commit 0d8ea4661c584fc095129f853d0d72b0fa09cda5)

Change-Id: Iad82178945a8f8c34943cc7573fd73d2b4decade